# packages/api-schema/openapi.yaml
openapi: 3.1.0
info:
  title: WISE Workbench API
  description: 'WISE Workbench backend: projects, datasets, mappings, case tables, norm versions, runs, backlogs, slice detail, traces, diagnostics, signal distributions, process flow and jobs. Errors are RFC 9457 problem details.'
  version: 0.1.0
paths:
...
  /projects/{projectId}:
    get:
      tags:
      - projects
      summary: Get Project
      operationId: getProject
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          title: Projectid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not Found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /projects/{projectId}/datasets:
    get:
      tags:
      - datasets
      summary: List Datasets
      operationId: listDatasets
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          title: Projectid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DatasetVersion'
                title: Response Listdatasets
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - datasets
      summary: Upload Dataset
      description: Multipart upload of CSV, Parquet or XES; returns the ingest job. The dataset id is in resultRef.
      operationId: uploadDataset
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          title: Projectid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_uploadDataset'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /projects/{projectId}/datasets/presets:
    get:
      tags:
      - datasets
      summary: List Presets
      description: Public logs with a known mapping and norm that can be loaded in one click; `available` says whether the files are on this machine.
      operationId: listPresets
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          title: Projectid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Preset'
                title: Response Listpresets
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /projects/{projectId}/datasets/presets/{presetId}:
    post:
      tags:
      - datasets
      summary: Load Preset
      description: 'Loads a public log preset: registers the configured file (not copied), builds the case table with the known mapping, imports the norm and scores a run, all in one job. The run id is in resultRef when done; every step reuses what the project already has.'
      operationId: loadPreset
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          title: Projectid
      - name: presetId
        in: path
        required: true
        schema:
          type: string
          title: Presetid
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not Found
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unprocessable Content
  /projects/{projectId}/datasets/{datasetId}:
    get:
      tags:
      - datasets
      summary: Get Dataset
      operationId: getDataset
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          title: Projectid
      - name: datasetId
        in: path
        required: true
        schema:
          type: string
          title: Datasetid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetVersion'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /projects/{projectId}/datasets/{datasetId}/preview:
    get:
      tags:
      - datasets
      summary: Preview Dataset
      operationId: previewDataset
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          title: Projectid
      - name: datasetId
        in: path
        required: true
        schema:
          type: string
          title: Datasetid
      - name: rows
        in: query
        required: false
        schema:
          type: integer
          maximum: 1000
          minimum: 1
          default: 50
          title: Rows
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Table'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /projects/{projectId}/case-tables:
    get:
      tags:
      - datasets
      summary: List Case Tables
      operationId: listCaseTables
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          title: Projectid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CaseTable'
                title: Response Listcasetables
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /projects/{projectId}/case-tables/{caseTableId}:
    get:
      tags:
      - datasets
      summary: Get Case Table
      operationId: getCaseTable
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          title: Projectid
      - name: caseTableId
        in: path
        required: true
        schema:
          type: string
          title: Casetableid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseTable'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /projects/{projectId}/case-tables/{caseTableId}/flow-types:
    get:
      tags:
      - datasets
      summary: Get Flow Types
      description: 'The detected flow types of a case table (from the mapping''s flow typing, else from the named attribute): counts, share, one process map each with stage groups, and a readiness headline per type. Each type carries the run scope that analyses it alone.'
      operationId: getFlowTypes
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          title: Projectid
      - name: caseTableId
        in: path
        required: true
        schema:
          type: string
          title: Casetableid
      - name: attribute
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 'case attribute holding the flow type (default: the mapping''s)'
          title: Attribute
        description: 'case attribute holding the flow type (default: the mapping''s)'
      - name: abstraction
        in: query
        required: false
        schema:
          type: number
          maximum: 1.0
          minimum: 0.0
          default: 0.05
          title: Abstraction
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlowTypes'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /projects/{projectId}/decisions/kinds:
    get:
      tags:
      - datasets
      summary: List Decision Kinds
      description: The decisions a reader can take on readiness items, with the parameters each accepts.
      operationId: listDecisionKinds
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          title: Projectid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DecisionKind'
                title: Response Listdecisionkinds
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /projects/{projectId}/decisions:
    get:
      tags:
      - datasets
      summary: List Decisions
      description: Decisions taken in the project (optionally those that touch one case table).
      operationId: listDecisions
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          title: Projectid
      - name: caseTableId
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Casetableid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Decision'
                title: Response Listdecisions
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /projects/{projectId}/case-tables/{caseTableId}/decisions/preview:
    post:
      tags:
      - datasets
      summary: Preview Decision
      description: Cases and events a decision would affect, before it is applied.
      operationId: previewDecision
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          title: Projectid
      - name: caseTableId
        in: path
        required: true
        schema:
          type: string
          title: Casetableid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecisionRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecisionPreviewOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /projects/{projectId}/case-tables/{caseTableId}/decisions:
    post:
      tags:
      - datasets
      summary: Apply Decision
      description: 'Apply a decision: it is stored as a versioned mapping decision (a child mapping), a new case table is built from it (job) and its readiness report is the re-evaluation.'
      operationId: applyDecision
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          title: Projectid
      - name: caseTableId
        in: path
        required: true
        schema:
          type: string
          title: Casetableid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecisionRequest'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecisionApplied'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Body_uploadDataset:
      properties:
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
      type: object
      required:
      - file
      title: Body_uploadDataset
    CaseTable:
      properties:
        id:
          type: string
          title: Id
        datasetId:
          type: string
          title: Datasetid
        mappingId:
          anyOf:
          - type: string
          - type: 'null'
          title: Mappingid
        cases:
          type: integer
          title: Cases
        events:
          anyOf:
          - type: integer
          - type: 'null'
          title: Events
        status:
          type: string
          enum:
          - building
          - ready
          - failed
          title: Status
        readiness:
          anyOf:
          - $ref: '#/components/schemas/Readiness'
          - type: 'null'
        activities:
          items:
            $ref: '#/components/schemas/ActivityCount'
          type: array
          title: Activities
        attributes:
          items:
            type: string
          type: array
          title: Attributes
        jobId:
          anyOf:
          - type: string
          - type: 'null'
          title: Jobid
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
        createdAt:
          type: string
          format: date-time
          title: Createdat
      type: object
      required:
      - id
      - datasetId
      - cases
      - status
      - createdAt
      title: CaseTable
...
    DatasetVersion:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        status:
          type: string
          enum:
          - ingesting
          - ready
          - failed
          title: Status
        contentHash:
          anyOf:
          - type: string
          - type: 'null'
          title: Contenthash
        events:
          anyOf:
          - type: integer
          - type: 'null'
          title: Events
        columns:
          items:
            $ref: '#/components/schemas/ColumnProfile'
          type: array
          title: Columns
        createdAt:
          type: string
          format: date-time
          title: Createdat
        jobId:
          anyOf:
          - type: string
          - type: 'null'
          title: Jobid
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
        sourceKind:
          anyOf:
          - type: string
          - type: 'null'
          title: Sourcekind
      type: object
      required:
      - id
      - name
      - status
      - createdAt
      title: DatasetVersion
    Decision:
      properties:
        id:
          type: string
          title: Id
        projectId:
          type: string
          title: Projectid
        caseTableId:
          type: string
          title: Casetableid
        kind:
          type: string
          title: Kind
        params:
          additionalProperties: true
          type: object
          title: Params
        readinessItem:
          type: string
          title: Readinessitem
        version:
          type: integer
          title: Version
        mappingId:
          type: string
          title: Mappingid
        resultCaseTableId:
          type: string
          title: Resultcasetableid
        preview:
          additionalProperties: true
          type: object
          title: Preview
        author:
          anyOf:
          - type: string
          - type: 'null'
          title: Author
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
        createdAt:
          type: string
          format: date-time
          title: Createdat
      type: object
      required:
      - id
      - projectId
      - caseTableId
      - kind
      - params
      - readinessItem
      - version
      - mappingId
      - resultCaseTableId
      - preview
      - createdAt
      title: Decision
    DecisionApplied:
      properties:
        decision:
          $ref: '#/components/schemas/Decision'
        caseTable:
          $ref: '#/components/schemas/CaseTable'
        job:
          $ref: '#/components/schemas/Job'
      type: object
      required:
      - decision
      - caseTable
      - job
      title: DecisionApplied
    DecisionKind:
      properties:
        kind:
          type: string
          title: Kind
        item:
          type: string
          title: Item
        params:
          items:
            type: string
          type: array
          title: Params
        label:
          type: string
          title: Label
      type: object
      required:
      - kind
      - item
      - params
      - label
      title: DecisionKind
    DecisionPreviewOut:
      properties:
        kind:
          type: string
          title: Kind
        params:
          additionalProperties: true
          type: object
          title: Params
        readinessItem:
          type: string
          title: Readinessitem
        label:
          type: string
          title: Label
        preview:
          additionalProperties: true
          type: object
          title: Preview
          description: cases, events, totalCases, totalEvents, detail
        caseTableId:
          type: string
          title: Casetableid
        version:
          type: integer
          title: Version
      type: object
      required:
      - kind
      - params
      - readinessItem
      - label
      - preview
      - caseTableId
      - version
      title: DecisionPreviewOut
    DecisionRequest:
      properties:
        kind:
          type: string
          title: Kind
          description: drop_outside_window | sentinel_as_missing | collapse_duplicates | day_precision | header_events | open_cases | zero_exposure | flow_type_assignment
        params:
          additionalProperties: true
          type: object
          title: Params
        author:
          anyOf:
          - type: string
          - type: 'null'
          title: Author
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
      type: object
      required:
      - kind
      title: DecisionRequest
...
    FlowTypes:
      properties:
        caseTableId:
          type: string
          title: Casetableid
        attribute:
          type: string
          title: Attribute
        source:
          type: string
          enum:
          - mapping
          - attribute
          title: Source
        cases:
          type: integer
          title: Cases
        windowEnd:
          anyOf:
          - type: string
          - type: 'null'
          title: Windowend
        caseNoun:
          anyOf:
          - type: string
          - type: 'null'
          title: Casenoun
        types:
          items:
            $ref: '#/components/schemas/FlowType'
          type: array
          title: Types
      type: object
      required:
      - caseTableId
      - attribute
      - source
      - cases
      - types
      title: FlowTypes
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    Job:
      properties:
        id:
          type: string
          title: Id
        kind:
          type: string
          title: Kind
        status:
          type: string
          enum:
          - queued
          - running
          - done
          - failed
          - cancelled
          title: Status
        progress:
          type: number
          title: Progress
          default: 0.0
        message:
          anyOf:
          - type: string
          - type: 'null'
          title: Message
        attempts:
          type: integer
          title: Attempts
          default: 0
        resultRef:
          anyOf:
          - type: string
          - type: 'null'
          title: Resultref
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
        createdAt:
          type: string
          format: date-time
          title: Createdat
        updatedAt:
          type: string
          format: date-time
          title: Updatedat
        projectId:
          anyOf:
          - type: string
          - type: 'null'
          title: Projectid
        cancelRequested:
          type: boolean
          title: Cancelrequested
          default: false
      type: object
      required:
      - id
      - kind
      - status
      - createdAt
      - updatedAt
      title: Job
...
    Preset:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        description:
          type: string
          title: Description
        available:
          type: boolean
          title: Available
        source:
          anyOf:
          - type: string
          - type: 'null'
          title: Source
          description: path of the log file on this machine
        norm:
          anyOf:
          - type: string
          - type: 'null'
          title: Norm
          description: path of the norm file on this machine
        mapping:
          $ref: '#/components/schemas/ColumnMapping'
        slicing:
          items:
            type: string
          type: array
          title: Slicing
        view:
          type: string
          title: View
        gamma:
          type: number
          title: Gamma
        minCases:
          type: integer
          title: Mincases
        process:
          type: string
          title: Process
      type: object
      required:
      - id
      - name
      - description
      - available
      - mapping
      - slicing
      - view
      - gamma
      - minCases
      - process
      title: Preset
    Problem:
      properties:
        type:
          type: string
          title: Type
          default: about:blank
        title:
          type: string
          title: Title
        status:
          type: integer
          title: Status
        detail:
          anyOf:
          - type: string
          - type: 'null'
          title: Detail
        instance:
          anyOf:
          - type: string
          - type: 'null'
          title: Instance
        code:
          anyOf:
          - type: string
          - type: 'null'
          title: Code
        errors:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Errors
      type: object
      required:
      - title
      - status
      title: Problem
    Project:
      properties:
        name:
          type: string
          minLength: 1
          title: Name
        process:
          anyOf:
          - type: string
          - type: 'null'
          title: Process
          description: knowledge pack id such as p2p or o2c
        question:
          anyOf:
          - type: string
          - type: 'null'
          title: Question
        id:
          type: string
          title: Id
        createdAt:
          type: string
          format: date-time
          title: Createdat
        latestRunId:
          anyOf:
          - type: string
          - type: 'null'
          title: Latestrunid
      type: object
      required:
      - name
      - id
      - createdAt
      title: Project
...
    Table:
      properties:
        columns:
          items:
            type: string
          type: array
          title: Columns
        rows:
          items:
            items: {}
            type: array
          type: array
          title: Rows
      type: object
      required:
      - columns
      - rows
      title: Table
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
servers:
- url: /api/v1